handles hyperlink columns and fields
